修正：MySQL长时间未通讯会丢失连接；改进：导入导出数据库时增加对数据库版本的检查；改进：页面上总是显示teleport服务器的时间（配合检查身份认证器的时间是否同步）；

+++ server/www/teleport/app/eom_app/controller/config.py
@@ -57,16 +57,17 @@ class IndexHandler(TPBaseAdminAuthHandler):
 
 class ExportDatabaseHandler(TPBaseAdminAuthHandler):
     def get(self):
-        now = time.localtime(time.time())
-        dt = '{:04d}{:02d}{:02d}-{:02d}{:02d}{:02d}'.format(now.tm_year, now.tm_mon, now.tm_mday, now.tm_hour, now.tm_min, now.tm_sec)
+        sql, ret = get_db().export_to_sql()
+        if ret:
+            now = time.localtime(time.time())
+            dt = '{:04d}{:02d}{:02d}-{:02d}{:02d}{:02d}'.format(now.tm_year, now.tm_mon, now.tm_mday, now.tm_hour, now.tm_min, now.tm_sec)
 
-        self.set_header('Content-Type', 'application/octet-stream')
-        self.set_header('Content-Disposition', 'attachment; filename=teleport-database-export-{}.sql'.format(dt))
+            self.set_header('Content-Type', 'application/octet-stream')
+            self.set_header('Content-Disposition', 'attachment; filename=teleport-database-export-{}.sql'.format(dt))
+            self.write(sql)
+        else:
+            self.write('<h1>错误</h1>导出数据时发生错误：{}'.format(sql))
 
-        sql = get_db().export_to_sql()
-
-        # self.write("分组ID, 操作系统, IP地址, 端口, 协议, 状态, 描述, 系统用户, 系统密码, 是否加密, 附加参数, 密钥ID, 认证类型\n".encode('gbk'))
-        self.write(sql)
         self.finish()
 
 
@@ -82,8 +83,7 @@ class ImportDatabaseHandler(TPBaseAdminAuthHandler):
         ret = dict()
         ret['code'] = 0
         ret['message'] = ''
-        # ret['data'] = {}
-        # ret['data']['msg'] = list()  # 记录跳过的行（格式不正确，或者数据重复等）
+
         sql_filename = ''
 
         try:
@@ -114,23 +114,39 @@ class ImportDatabaseHandler(TPBaseAdminAuthHandler):
                 ret['message'] = 'upload sql file is not utf8 encode.'
                 return self.write(json.dumps(ret).encode('utf8'))
 
+            db_ver_checked = False
             with open(sql_filename, encoding=file_encode) as f:
                 db = get_db()
                 sql = []
                 lines = f.readlines()
                 for line in lines:
                     line = line.strip('\r\n')
+                    if line.startswith('-- DATABASE VERSION '):
+                        x = line.split(' ')
+                        if len(x) != 4:
+                            ret['code'] = -1
+                            ret['message'] = 'SQL文件格式错误，无法解析数据库版本'
+                            return self.write(json.dumps(ret).encode('utf8'))
+                        db_ver_sql = int(x[3].strip())
+                        if db.DB_VERSION != db_ver_sql:
+                            ret['code'] = -1
+                            ret['message'] = 'SQL文件数据库版本为 {}，当前数据版本为 {}，不允许导入！'.format(db_ver_sql, db.DB_VERSION)
+                            return self.write(json.dumps(ret).encode('utf8'))
+                        db_ver_checked = True
+                        continue
+
+                    if not db_ver_checked:
+                        continue
+
                     if line .startswith('TRUNCATE TABLE '):
                         x = line.split(' ', 2)
-                        _table_name = '`{}{}`'.format(get_db().table_prefix, x[2][1:-2])
+                        _table_name = '`{}{}`'.format(db.table_prefix, x[2][1:-2])
                         if db.db_type == db.DB_TYPE_MYSQL:
                             x[2] = _table_name
                             line = ' '.join(x)
                             line += ';'
                             sql.append(line)
                         elif db.db_type == db.DB_TYPE_SQLITE:
-                            #delete from TableName;  //清空数据
-                            # update sqlite_sequence SET seq = 0 where name ='TableName';//自增长ID为0
                             sql.append('DELETE FROM {};'.format(_table_name))
                             sql.append('UPDATE `sqlite_sequence` SET `seq`=0 WHERE `name`="{}";'.format(_table_name[1:-1]))
 
@@ -139,21 +155,18 @@ class ImportDatabaseHandler(TPBaseAdminAuthHandler):
                         _table_name = '`{}{}`'.format(db.table_prefix, x[2][1:-1])
                         x[2] = _table_name
                         line = ' '.join(x)
-                        # print(line)
                         sql.append(line)
+
+                if not db_ver_checked:
+                    ret['code'] = -1
+                    ret['message'] = 'SQL文件格式错误，未能确定数据库版本'
+                    return self.write(json.dumps(ret).encode('utf8'))
 
                 db_ret = db.transaction(sql)
                 if not db_ret:
                     ret['code'] = -1
                     ret['message'] = 'SQL语句执行出错'
                     return self.write(json.dumps(ret).encode('utf8'))
-
-                # for line in sql:
-                #     db_ret = get_db().exec(line)
-                #     if not db_ret:
-                #         ret['code'] = -1
-                #         ret['message'] = 'SQL语句执行出错: {}'.format(line)
-                #         return self.write(json.dumps(ret).encode('utf8'))
 
             ret['code'] = 0
             return self.write(json.dumps(ret).encode('utf8'))
